Update UPFDeployment pods on a config-map update. (#15)
Initial implementation of configuration changes in UPF.
[101](https://github.com/nephio-project/nephio/issues/101)
In case of config-map update, we update an Annotation in
Spec.Template.Annotations that trigger a deployment rolling update.
For the time being, we update the existing config map rather than
creating a new one. The best practice described below is a subject for
R2 or further releases.

Per John:
The best practice would be to create separate ConfigMap as they change.
Then, you update the ConfigMap that the deployment volume refers to,
which automatically does the rolling update.
By doing this, we ensure that the old replica set doesn't pick up the
new ConfigMap. So, if the old Pod restarts, it still has the old config.